Limit request frequency for user model to 2 seconds
The landscapes route requests a user model from the user service.
When e.g. the user-service is not available, the returned Promise
of the model-hook is rejected. This re-triggers the model-hook and
thereby can lead to a UI-freezing loop requesting the user data.

Solution: Limit the fetching of user data to at most every 2 seconds.