Fix recursive class communication rendering
The computation for class communication was not correct for recursive
class communication. For the special case that there is only one
request for the same class, the line thickness was calculated as NaN.
One cause was that recursive requests were omitted in the calculation.
This commit takes recursive requests into account and fixes the
visualization of recursive class communication calls.