Fix: Replace components only when box layout changes
The box layout of an application is stored in maps.
Up until now, the map references were compared to determine
if the layout changed. This is not correct, because the
references are updated no matter if the layout changed or not.
Instead, we now compare the size of the maps.
This is efficient and should be correct a change in layout without
interacting with components may only occur when new components
are discovered by the dynamic code instrumentation.