Rename "Vision Impairment' to 'Blue' color scheme
The name "vision impairment" could be seen as misleading, as
there are many types of vision impairment, and this color scheme
was not systematically tested for any of them.
In addition to the name change, the color of classes is changed
to increase its contrast with the components and labels.