Replace DrawableClassCommunication with AggregatedClassCommunication
The DrawableClassCommunication was a class that was used to represent
the aggregated calls of one method. This, however, was not the original
meaning of "drawable". Hereby, drawable means that the object can be
added as a datamodel to a mesh and can be rendered without further
computations necessary.
This commit introduces aggregatedMethodCalls, representing all calls
to a specific method and aggregatedClassCommunication representing all
aggregatedMethodCalls between two classes (in both directions).