Limit device pixel ratio to 2
Some devices, especially mobile devices, have a device pixel ratio
of 3 or higher. This means that 9x more pixels need to be rendered
than on a normal screen. This can cause performance issues.
Since a pixel ratio of 2 (which Apple would call "Retina") is more
than enough from a quality perspective, we limit the pixel ratio
to 2.